requirements.txt
Flask
tensorflow
numpy
matplotlib
pillow